feat(src/library/native_compiler): expose more compiler config
Specifically we add the ability to choose a place to store the generated code (native.store_code),
a flag for setting the executable name (native.binary), and full support for setting the compiler
name with (native.cc). We now also generate a temporary unique file during compilation to fix an
issue in testing.